requirements.txt
Selenium-Screenshot==2.0.0
webdriver-manager==3.8.5